DB migration and initial refactor
