+++ src/fides/api/schemas/user.py
@@ -145,3 +145,10 @@ class DisabledReason(Enum):
     """Reasons for why a user is disabled"""
 
     pending_invite = "pending_invite"
+
+
+class LoginMethod(str, Enum):
+    """Methods for logging in"""
+
+    username_password = "username_password"
+    sso = "sso"
